requirements.txt
langchain-huggingface
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.20
langchain-openai>=0.0.6
langchain-cohere>=0.1.0
langchain-unstructured>=0.1.0

faiss-cpu>=1.7.4
numpy>=1.24.0

PyMuPDF>=1.23.0
unstructured[all-docs]>=0.10.0
requests>=2.31.0
flask
mysql-connector-python
camelot-py>=1.0.9

google-search-results>=2.4.2
python-dotenv>=1.0.0
langsmith>=0.1.0